+++ client/src/App.css
@@ -1,5 +1,8 @@
 body {
   color: #ff66b3;
+  /* background-image: url("https://images.unsplash.com/photo-1600636214973-8ccfde1f6a67?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=987&q=80");
+  background-repeat: no-repeat;
+  background-size: cover; */
 }
 
 /* ul {
@@ -41,7 +44,7 @@ div.middle {
 
 .blogForm {
   margin: 25px;
-  background-color: #dfc3a5;
+  background-color: #bbe6e4;
   border: 2px solid black;
   border-radius: 10px;
   opacity: 0.75;
@@ -81,10 +84,10 @@ div.cardPostList {
 
 }
 
-span {
-  
-  
+li {
+  list-style: none;
 };
+
 /*
  .button {
   transition-duration: 0.4s;
@@ -94,3 +97,9 @@ span {
   background-color: #4CAF50; /* Green */
 /* color: white;
 };  */
+/*
+Opacity tricks:
+background: rgba(76, 175, 80, 0.3) comment out? Green background with 30% opacity comment out?
+Gaby's example with hex code instead
+background-color: fade(#5A767B, 50%);
+*/
