+++ client/src/Pages/Korea.js
@@ -1,11 +1,39 @@
-import React from 'react';
+import IndividualPost from "../components/IndividualPost";
+import Form from "../components/Form";
+import KoreaPostList from "../components/KoreaPostList";
+import { useState } from "react";
 
 const Korea = () => {
-return (
-	<div>
-	<h1>Korea Ventures heeeeere</h1>
-	</div>
-);
+	const [selected, setSelected] = useState(null);
+	//const [posts, setPosts] = useState([]);
+  
+	function passingSelected(post) {
+	  //console.log('hello')
+	  setSelected(post);
+	  console.log("post dallas check", post);
+	}
+  
+	let switchDisplayView;
+  
+	if (selected != null) {
+	  switchDisplayView = (
+		<div>
+		  <IndividualPost post={selected} />
+		  <br />
+  
+		  <button onClick={() => setSelected(null)}>Go Back</button>
+		</div>
+	  );
+	} else {
+	  switchDisplayView = (
+		<div>
+		  <Form header={"Korea Ventures Heeere"} location={"Gunsan, South Korea"}/>
+		  <KoreaPostList passingSelected={passingSelected} />
+		</div>
+	  );
+	}
+  
+	return <div>{switchDisplayView}</div>;
 };
 
 export default Korea;
