+++ client/src/components/DallasPostList.js
@@ -1,13 +1,14 @@
 import { useEffect, useState } from "react";
 //import IndividualPost from "./IndividualPost";
 //import EditForm from "./EditForm";
+import Form from "./Form";
 
 function DallasPostList(props) {
   const [posts, setPosts] = useState([]);
 
   //let selected = props;
 
-  const [editingPostId, setEditingPostId] = useState([]);
+  const [editingPostId, setEditingPostId] = useState(null);
   const [user, setUser] = useState(undefined);
 
   useEffect(() => {
@@ -48,21 +49,28 @@ function DallasPostList(props) {
     props.savePost(data);
   };
 
-  const updatePost = (savePost) => {
+  const updatePost = (updatedPost) => {
     setPosts((posts) => {
-      const newPost = [];
+      const newPostList = [];
       for (let post of posts) {
-        if (post.id === savePost.id) {
-          newPost.push(savePost);
+        if (post.id === updatedPost.id) {
+          newPostList.push(updatedPost);
         } else {
-          newPost.push(post);
+          newPostList.push(post);
         }
       }
-      return newPost;
+      return newPostList;
     });
     //this is meant to close the form
     setEditingPostId(null);
   };
+
+//a function to grab the post id of the student that we want to edit
+const onEdit = (post) => {
+  const editingId = post.id;
+  console.log(editingId);
+  setEditingPostId(editingId);
+};
 
   //logic for deleting an existing post by id
   const onDelete = async (id) => {
@@ -93,10 +101,20 @@ function DallasPostList(props) {
     props.passingSelected(post);
   };
 
+  const addPost = (newPost) => {
+    //console.log(newPost);
+    setPosts((posts) => [...posts, newPost]);
+  };
+
   //console.log(posts);
   return (
     <div>
-      {posts.map((post) => (
+      {posts.map((post) => {
+        if (post.id === editingPostId) {
+          console.log("dpList prop check", post);
+          return <Form header={"Editing Mode"} location={"Dallas, TX"} initialPost={post} addPost={addPost}/>;
+        } else {
+          return (
         <div key={post.id} className="cardPostList" style={{ width: "20rem" }}>
           <img
             className="card-img-top"
@@ -125,8 +143,11 @@ function DallasPostList(props) {
               View
             </button>
             {user && (
-              <li>
-                <button onClick={handleSubmit}>Edit</button>
+              <>
+                <button onClick={(e) => {
+                  e.preventDefault();
+                  onEdit(post)
+                }}>Edit</button>
                 <button
                   type="button"
                   onClick={() => {
@@ -135,11 +156,13 @@ function DallasPostList(props) {
                 >
                   Delete
                 </button>
-              </li>
+                </>
             )}
           </div>
         </div>
-      ))}
+          )
+                }
+              })}
     </div>
   );
 }
